Refactor docstrings across multiple modules to improve clarity and consistency with Google style
- Updated parameter descriptions to use "Args" instead of "Parameters" for consistency.
- Improved clarity of function descriptions and return types.
- Standardized formatting of docstrings to follow a consistent style.
- Enhanced readability by ensuring uniformity in the presentation of arguments and return values.
- Adjusted example sections for clarity in usage.

+++ rubix/telescope/apertures.py
@@ -1,7 +1,6 @@
 """This class defines the aperture mask for the observation of a galaxy."""
 
 import jax.numpy as jnp
-import numpy as np
 from beartype import beartype as typechecker
 from jaxtyping import Array, Float, jaxtyped
 
@@ -9,7 +8,7 @@ __all__ = ["HEXAGONAL_APERTURE", "SQUARE_APERTURE", "CIRCULAR_APERTURE"]
 
 
 @jaxtyped(typechecker=typechecker)
-def HEXAGONAL_APERTURE(sbin: np.int64) -> Float[Array, "..."]:
+def HEXAGONAL_APERTURE(sbin: int) -> Float[Array, "..."]:
     """
     Creates a hexagonal aperture mask for the observation of a galaxy.
 
@@ -42,9 +41,9 @@ def HEXAGONAL_APERTURE(sbin: np.int64) -> Float[Array, "..."]:
 
 
 @jaxtyped(typechecker=typechecker)
-def SQUARE_APERTURE(sbin: np.int64) -> Float[Array, "..."]:
-    """Creates a square aperture mask for the observation of a galaxy.
-
+def SQUARE_APERTURE(sbin: int) -> Float[Array, "..."]:
+    """
+    Creates a square aperture mask for the observation of a galaxy.
     Args:
         sbin (int): The size of the spatial bin in each direction for the aperture mask.
 
@@ -57,9 +56,9 @@ def SQUARE_APERTURE(sbin: np.int64) -> Float[Array, "..."]:
 
 
 @jaxtyped(typechecker=typechecker)
-def CIRCULAR_APERTURE(sbin: np.int64) -> Float[Array, "..."]:
-    """Creates a circular aperture mask for the observation of a galaxy.
-
+def CIRCULAR_APERTURE(sbin: int) -> Float[Array, "..."]:
+    """
+    Creates a circular aperture mask for the observation of a galaxy.
     Args:
         sbin (int): The size of the spatial bin in each direction for the aperture mask.
